version 2.0 initial commit

+++ scripts.d/500_start_iperf.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-
 DESCRIPTION="Start an iperf server"
+SCRIPT_TYPE="parallel"
+
+which iperf &> /dev/null
+if [ $? == 1 ]; then
+    write_log "ERROR: iperf not installed"
+    exit -1
+fi
 
 # Put your stuff here
 sudo pkill iperf	# make sure it's not already running
